linted macro tests
